requirements.txt
plotly == 4.14.3
plotly.express == 0.4.1
numpy == 1.20.0
pandas == 1.2.0
pydeck == 0.5.0
streamlit == 0.79.0
ptitprince == 0.2.5
matplotlib == 3.3.4
japanize-matplotlib == 1.1.3
scikit-learn == 0.24.1
scipy == 1.5.4
shap == 0.39.0
dtreeviz == 1.1.4
graphviz == 0.16
seaborn == 0.11.1
sklearn == 0.0
factor-analyzer == 0.3.2
nlplot == 1.2.0
statsmodels == 0.12.2
Cython == 0.29.22
sklearn == 0.0
pillow == 8.0.1
